fix: Correct bill_id foreign key data type in bill_payments migration
Changed bill_id from unsignedBigInteger to unsignedInteger to match
bills table's increments('id') which creates UNSIGNED INTEGER.

Error: SQLSTATE[HY000]: General error: 3780 Referencing column 'bill_id'
and referenced column 'id' in foreign key constraint are incompatible.

Root cause: bills.id uses increments() (32-bit), but bill_payments.bill_id
used unsignedBigInteger (64-bit).

+++ database/migrations/2025_11_11_150004_create_bill_payments_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->string('unique_hash')->nullable();
 
             // Relationships
-            $table->unsignedBigInteger('bill_id')->nullable();
+            $table->unsignedInteger('bill_id')->nullable();
             $table->foreign('bill_id')->references('id')->on('bills')->onDelete('cascade');
             $table->integer('company_id')->unsigned()->nullable();
             $table->foreign('company_id')->references('id')->on('companies')->onDelete('restrict');
